add deps for tfhub 0.3 and udpate code to make it work
