+++ python/requirements.txt
@@ -3,3 +3,4 @@ keras==2.2.4
 numpy==1.15.1
 six==1.11.0
 tf-nightly-2.0-preview==2.0.0.dev20190301
+tensorflow-hub==0.3.0
