+++ python/tensorflowjs/converters/keras_tfjs_loader_test.py
@@ -43,7 +43,7 @@ class LoadKerasModelTest(tf.test.TestCase):
   def tearDown(self):
     if os.path.isdir(self._tmp_dir):
       shutil.rmtree(self._tmp_dir)
-    tf.reset_default_graph()
+    tf.compat.v1.reset_default_graph()
     super(LoadKerasModelTest, self).tearDown()
 
   def _saveKerasModelForTest(self, path):
@@ -59,13 +59,13 @@ class LoadKerasModelTest(tf.test.TestCase):
 
   def testLoadKerasModelAndWeights(self):
     """Test loading of model and its weights."""
-    # Use separate tf.Graph and tf.Session contexts to prevent name collision.
-    with tf.Graph().as_default(), tf.Session():
+    # Use separate tf.Graph and tf.compat.v1.Session contexts to prevent name collision.
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           os.path.join(tfjs_path, 'model.json'))
 
@@ -81,8 +81,8 @@ class LoadKerasModelTest(tf.test.TestCase):
 
   def testDeserializeKerasModelTopologyOnlyFromBytesIO(self):
     """Test loading of model (only topology) from a BytesIO object."""
-    # Use separate tf.Graph and tf.Session contexts to prevent name collision.
-    with tf.Graph().as_default(), tf.Session():
+    # Use separate tf.Graph and tf.compat.v1.Session contexts to prevent name collision.
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
 
@@ -94,7 +94,7 @@ class LoadKerasModelTest(tf.test.TestCase):
     buff_writer.flush()
     buff_writer.seek(0)
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.deserialize_keras_model(buff.read())
 
       # The two model JSONs should match exactly.
@@ -102,8 +102,8 @@ class LoadKerasModelTest(tf.test.TestCase):
 
   def testDeserializeKerasModelTopologyOnlyFromJSONDict(self):
     """Test loading of model (only topology) from a JSON Dict."""
-    # Use separate tf.Graph and tf.Session contexts to prevent name collision.
-    with tf.Graph().as_default(), tf.Session():
+    # Use separate tf.Graph and tf.compat.v1.Session contexts to prevent name collision.
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
 
@@ -111,7 +111,7 @@ class LoadKerasModelTest(tf.test.TestCase):
     with open(os.path.join(tfjs_path, 'model.json'), 'rt') as f:
       config_json = json.load(f)
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.deserialize_keras_model(config_json)
 
       # The two model JSONs should match exactly.
@@ -119,8 +119,8 @@ class LoadKerasModelTest(tf.test.TestCase):
 
   def testDeserializeKerasModelTopologyAndWeightsFromBuffers(self):
     """Test loading of model and its weights from buffers."""
-    # Use separate tf.Graph and tf.Session contexts to prevent name collision.
-    with tf.Graph().as_default(), tf.Session():
+    # Use separate tf.Graph and tf.compat.v1.Session contexts to prevent name collision.
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
@@ -135,7 +135,7 @@ class LoadKerasModelTest(tf.test.TestCase):
       with open(path, 'rb') as f:
         weight_buffers.append(f.read())
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.deserialize_keras_model(
           json_buff, weight_data=weight_buffers)
 
@@ -151,8 +151,8 @@ class LoadKerasModelTest(tf.test.TestCase):
 
   def testDeserializeKerasModelTopologyAndWeightsFromFileObjects(self):
     """Test loading of model and its weights using file objects."""
-    # Use separate tf.Graph and tf.Session contexts to prevent name collision.
-    with tf.Graph().as_default(), tf.Session():
+    # Use separate tf.Graph and tf.compat.v1.Session contexts to prevent name collision.
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
@@ -163,7 +163,7 @@ class LoadKerasModelTest(tf.test.TestCase):
     weight_paths = sorted(glob.glob(os.path.join(tfjs_path, 'group*')))
     weight_files = [open(path, 'rb') for path in weight_paths]
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.deserialize_keras_model(
           json_file, weight_files)
 
@@ -182,13 +182,13 @@ class LoadKerasModelTest(tf.test.TestCase):
       f.close()
 
   def testLoadKerasModelWithCurrentWorkingDirectoryRelativePath(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
 
     os.chdir(tfjs_path)
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       # Use a relative path under the current working directory.
       model2 = keras_tfjs_loader.load_keras_model('model.json')
 
@@ -204,12 +204,12 @@ class LoadKerasModelTest(tf.test.TestCase):
 
   def testLoadKerasModelWithoutWeights(self):
     """Test loading of model topology only, without loading weight values."""
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           os.path.join(tfjs_path, 'model.json'), load_weights=False)
       model2_weight_values = model2.get_weights()
@@ -227,7 +227,7 @@ class LoadKerasModelTest(tf.test.TestCase):
       self.assertEqual(model1.to_json(), model2.to_json())
 
   def testLoadKerasModelFromNonDefaultWeightsPathWorks(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
@@ -238,7 +238,7 @@ class LoadKerasModelTest(tf.test.TestCase):
       new_model_json_path = os.path.join(model_json_dir, 'model.json')
       shutil.move(os.path.join(tfjs_path, 'model.json'), new_model_json_path)
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           new_model_json_path, weights_path_prefix=tfjs_path)
 
@@ -253,7 +253,7 @@ class LoadKerasModelTest(tf.test.TestCase):
 
   def testLoadKerasModelWithUniqueNameScopeInTheSameGraphContext(self):
     """Test enabling unique name scope during model loading."""
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
@@ -275,7 +275,7 @@ class LoadKerasModelTest(tf.test.TestCase):
         self.assertTrue(name2.endswith(name1))
 
   def testLoadKerasModelFromDataBuffers(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       model1 = self._saveKerasModelForTest(tfjs_path)
       model1_weight_values = model1.get_weights()
@@ -290,7 +290,7 @@ class LoadKerasModelTest(tf.test.TestCase):
           with open(os.path.join(tfjs_path, path), 'rb') as f:
             data_buffer += f.read()
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           os.path.join(tfjs_path, 'model.json'),
           weights_data_buffers=data_buffers)
@@ -305,7 +305,7 @@ class LoadKerasModelTest(tf.test.TestCase):
       self.assertEqual(model1.to_json(), model2.to_json())
 
   def testLoadNestedKerasModel(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       inner_model = keras.Sequential([
           keras.layers.Dense(4, input_shape=[3], activation='relu'),
           keras.layers.Dense(3, activation='tanh')])
@@ -319,13 +319,13 @@ class LoadKerasModelTest(tf.test.TestCase):
       save_dir = os.path.join(self._tmp_dir, 'nested_model')
       keras_h5_conversion.save_keras_model(outer_model, save_dir)
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           os.path.join(save_dir, 'model.json'), use_unique_name_scope=True)
       self.assertAllClose(predict_out, model2.predict(x))
 
   def testLoadNestedTfKerasModel(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       inner_model = tf.keras.Sequential([
           tf.keras.layers.Dense(4, input_shape=[3], activation='relu'),
           tf.keras.layers.Dense(3, activation='tanh')])
@@ -340,13 +340,13 @@ class LoadKerasModelTest(tf.test.TestCase):
       save_dir = os.path.join(self._tmp_dir, 'nested_model')
       keras_h5_conversion.save_keras_model(outer_model, save_dir)
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           os.path.join(save_dir, 'model.json'), use_unique_name_scope=True)
       self.assertAllClose(predict_out, model2.predict(x))
 
   def testLoadKerasModeFromNonexistentWeightsPathRaisesError(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       self._saveKerasModelForTest(tfjs_path)
       with self.assertRaises(ValueError):
@@ -355,7 +355,7 @@ class LoadKerasModelTest(tf.test.TestCase):
             weights_path_prefix=os.path.join(self._tmp_dir, 'nonexistent'))
 
   def testUsingBothWeightsDataBuffersAndWeightsPathPrefixRaisesError(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       self._saveKerasModelForTest(tfjs_path)
 
@@ -365,7 +365,7 @@ class LoadKerasModelTest(tf.test.TestCase):
             weights_data_buffers=[b'foo'], weights_path_prefix='bar')
 
   def testInvalidJSONRaisesError(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       tfjs_path = os.path.join(self._tmp_dir, 'model_for_test')
       self._saveKerasModelForTest(tfjs_path)
 
@@ -383,7 +383,7 @@ class LoadKerasModelTest(tf.test.TestCase):
             weights_data_buffers=[b'foo'], weights_path_prefix='bar')
 
   def testLoadFunctionalKerasModel(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       input1 = keras.Input([4])
       x1 = keras.layers.Dense(2, activation='relu')(input1)
       x1 = keras.layers.BatchNormalization()(x1)
@@ -405,14 +405,14 @@ class LoadKerasModelTest(tf.test.TestCase):
       save_dir = os.path.join(self._tmp_dir, 'functional_model')
       keras_h5_conversion.save_keras_model(model, save_dir)
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           os.path.join(save_dir, 'model.json'))
       self.assertAllClose(
           predict_out, model2.predict([input1_val, input2_val]))
 
   def testLoadFunctionalTfKerasModel(self):
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       input1 = tf.keras.Input([4])
       x1 = tf.keras.layers.Dense(2, activation='relu')(input1)
       x1 = tf.keras.layers.BatchNormalization()(x1)
@@ -434,7 +434,7 @@ class LoadKerasModelTest(tf.test.TestCase):
       save_dir = os.path.join(self._tmp_dir, 'functional_model')
       keras_h5_conversion.save_keras_model(model, save_dir)
 
-    with tf.Graph().as_default(), tf.Session():
+    with tf.Graph().as_default(), tf.compat.v1.Session():
       model2 = keras_tfjs_loader.load_keras_model(
           os.path.join(save_dir, 'model.json'))
       self.assertAllClose(
